streamlit/requirements1.txt
absl-py @ file:///croot/absl-py_1686852429912/work
aiobotocore @ file:///croot/aiobotocore_1701291493089/work
aiohttp @ file:///croot/aiohttp_1701112538292/work
aioitertools @ file:///tmp/build/80754af9/aioitertools_1607109665762/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
altair==5.2.0
anyio @ file:///work/ci_py311/anyio_1676823771847/work/dist
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto @ file:///work/ci_py311/asn1crypto_1676838975389/work
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
botocore @ file:///croot/botocore_1701286451219/work
Bottleneck @ file:///work/ci_py311/bottleneck_1676827011612/work
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
cachetools==5.3.2
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer==3.3.2
click==8.1.7
cloudpickle @ file:///croot/cloudpickle_1683040006038/work
comm==0.2.1
cryptography @ file:///croot/cryptography_1702070282333/work
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastapi @ file:///croot/fastapi_1693295406697/work
fastjsonschema==2.19.1
filelock @ file:///croot/filelock_1700591183607/work
fqdn==1.5.1
frozenlist @ file:///croot/frozenlist_1698702560391/work
fsspec @ file:///croot/fsspec_1701286474621/work
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
ipykernel==6.28.0
ipython==8.19.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath @ file:///croot/jmespath_1700144569655/work
joblib @ file:///croot/joblib_1685113087166/work
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.6.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
multidict @ file:///croot/multidict_1701096859099/work
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1704311704800/work/dist/numpy-1.26.3-cp311-cp311-linux_x86_64.whl#sha256=10a078151ecec16bafb535f7487635217625fa06536dec8509e514648c78d626
oauth2client==4.1.3
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1695798841041/work
pydeck==0.8.1b0
PyDrive==1.3.1
Pygments==2.17.2
PyJWT @ file:///work/ci_py311/pyjwt_1676827385359/work
pymssql==2.2.11
pyodbc @ file:///croot/pyodbc_1705431351638/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.2
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///croot/python-dateutil_1694417990566/work
python-decouple==3.8
python-json-logger==2.0.7
pytimeparse @ file:///work/ci_py311/pytimeparse_1676838679391/work
pytz @ file:///croot/pytz_1695131579487/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==25.1.2
referencing==0.32.0
requests @ file:///croot/requests_1690400202158/work
retrying @ file:///Users/ktietz/demo/mc3/conda-bld/retrying_1629465456590/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
s3fs @ file:///croot/s3fs_1701294169021/work
scikit-learn @ file:///croot/scikit-learn_1684954695550/work
scipy @ file:///croot/scipy_1701295040508/work/dist/scipy-1.11.4-cp311-cp311-linux_x86_64.whl#sha256=4a1942ba48a330f141b1056df960d77f9be4cee3e2d999d754e791da201edffd
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python @ file:///croot/snowflake-connector-python_1701985259118/work
snowflake-ml-python @ file:///croot/snowflake-ml-python_1706040862704/work/bazel-bin/snowflake/ml/snowflake_ml_python-1.2.0-py3-none-any.whl#sha256=23cc5a7d5a43655adb9d66d1e2f6a077218abcde83ca1977524bc7712acede68
snowflake-snowpark-python==1.12.0
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1623949099177/work
soupsieve==2.5
sqlparse @ file:///croot/sqlparse_1690901927333/work
st-pages==0.4.5
stack-data==0.6.3
starlette @ file:///croot/starlette-recipe_1692980413034/work
streamlit==1.30.0
streamlit-aggrid==0.3.4.post3
tenacity==8.2.3
terminado==0.18.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tinycss2==1.2.1
toml==0.10.2
tomlkit @ file:///work/ci_py311/tomlkit_1676823218885/work
toolz==0.12.0
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.14
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wrapt @ file:///work/ci_py311/wrapt_1676827586170/work
xgboost==1.7.3
yarl @ file:///croot/yarl_1701105127787/work
zipp @ file:///croot/zipp_1704206909481/work
